Create seperate files for 'addTestItem' and 'testList' components.
